# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ShopifyProductVariant {
  id: string;
  title: string;
  availableForSale: boolean;
  priceAmount: string;
  currencyCode: string;
}

export interface ShopifyProduct {
  handle: string;
  title: string;
  imageUrl?: string;
  imageAlt?: string;
  variants: ShopifyProductVariant[];
}

export interface ShopifyProductCatalogResponse {
  products: ShopifyProduct[];
}

export interface ShopifyCheckoutLine {
  /** @minLength 1 */
  handle: string;
  /** @minLength 1 */
  variantId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}

export interface CreateShopifyCheckoutRequest {
  /**
     * @minItems 1
     * @maxItems 50
     */
  lines: ShopifyCheckoutLine[];
}

export interface ShopifyCheckoutRedirectResponse {
  checkoutUrl: string;
}

export interface ApiError {
  error: string;
}