# lib/api-zod/src/generated/types/createShopifyCheckoutRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyCheckoutLine } from './shopifyCheckoutLine';

export interface CreateShopifyCheckoutRequest {
  /**
     * @minItems 1
     * @maxItems 50
     */
  lines: ShopifyCheckoutLine[];
}