# lib/api-zod/src/generated/types/shopifyCheckoutLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopifyCheckoutLine {
  /** @minLength 1 */
  handle: string;
  /** @minLength 1 */
  variantId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}