# lib/api-zod/src/generated/types/shopifyProduct.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyProductVariant } from './shopifyProductVariant';

export interface ShopifyProduct {
  handle: string;
  title: string;
  imageUrl?: string;
  imageAlt?: string;
  variants: ShopifyProductVariant[];
}