# lib/api-zod/src/generated/types/shopifyProductCatalogResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyProduct } from './shopifyProduct';

export interface ShopifyProductCatalogResponse {
  products: ShopifyProduct[];
}